Add double support in json

+++ CMakeProject1/BusManager/json.h
@@ -9,7 +9,7 @@
 namespace Json {
     class Node : std::variant<std::vector<Node>,
         std::map<std::string, Node>,
-        int,
+        double,
         std::string> {
     public:
         using variant::variant;
@@ -20,8 +20,8 @@ namespace Json {
         const auto& AsMap() const {
             return std::get<std::map<std::string, Node>>(*this);
         }
-        int AsInt() const {
-            return std::get<int>(*this);
+        double AsDouble() const {
+            return std::get<double>(*this);
         }
         const auto& AsString() const {
             return std::get<std::string>(*this);
